--- a/test/import-blacklist_test.go
+++ b/test/import-blacklist_test.go
@@ -14,3 +14,13 @@ func TestImportsBlacklist(t *testing.T) {
 		Arguments: args,
 	})
 }
+
+func BenchmarkImportsBlacklist(b *testing.B) {
+	args := []interface{}{"crypto/md5", "crypto/sha1"}
+	var t *testing.T
+	for i := 0; i <= b.N; i++ {
+		testRule(t, "imports-blacklist", &rule.ImportsBlacklistRule{}, &lint.RuleConfig{
+			Arguments: args,
+		})
+	}
+}