Refactoring of `import-blacklist` to simplify code by removing unnecessary call to `ast.Walk` + other minor modifications.

Current vs new benchmarks:

```
$ benchdiff old.txt new.txt
benchmark                       old ns/op     new ns/op     delta
BenchmarkImportsBlacklist-4     114217        111158        -2.68%

benchmark                       old allocs     new allocs     delta
BenchmarkImportsBlacklist-4     455            448            -1.54%

benchmark                       old bytes     new bytes     delta
BenchmarkImportsBlacklist-4     38589         38265         -0.84%
```